Removed `mqtt_communication` module; replaced direct calls with `mqtt_service` abstraction, updated headers, and refactored key workflows for consistency.
