+++ Hardware/ESPIDF-Server/main/ir_service/ir_remote.cpp
@@ -1,0 +1,5 @@
+//
+// Created by angel on 22.01.26 г..
+//
+
+#include "ir_remote.h"
