+++ Hardware/ESPIDF-Server/main/ir_service/ir_remote.h
@@ -1,0 +1,13 @@
+//
+// Created by angel on 22.01.26 г..
+//
+
+#ifndef ESPIDF_IR_REMOTE_H
+#define ESPIDF_IR_REMOTE_H
+
+
+class ir_remote {
+};
+
+
+#endif //ESPIDF_IR_REMOTE_H
