+++ Hardware/ESPIDF-Server/main/wifi_service/wifi_control.h
@@ -4,8 +4,8 @@ extern "C" {
 #endif
 
     void wifi_init_setup();
-    void wifi_init_sta(char* ssid, char* password);
-    void change_wifi(char* new_ssid, char* new_password);
+    void wifi_init_sta(char* ssid, const char* password);
+    void change_wifi(char* new_ssid, const char* new_password);
 
     extern bool is_user_initiated_disconnect;
     extern int IS_WIFI_CONNECTED;
